TS-ify use-load-more and use TS and defineComponent in layouts and pages
